@@ -160,7 +160,7 @@ build-backend = "hatchling.build"
 fix = true
 include = ["*.py"]
 src = ["spectrafit"]
-target-version = "py38"
+target-version = "py39"
 
 [tool.ruff.lint]
 select = ["ALL"]
